requirements.txt
# requirements.txt

aiogram>=3.13.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary

# HTTP Client
aiohttp>=3.9.0

# Config
python-dotenv>=1.0.0

# Logging
structlog>=24.1.0

# Async file operations
aiofiles>=23.2.1

alembic